requirements.txt
cupy_cuda12x==14.0.1
matplotlib==3.10.9
numpy==2.4.4
pandas==3.0.2
Pillow==12.2.0
setuptools==75.1.0